+++ tests/envinterp/__init__.py
@@ -1,0 +1,10 @@
+def setup_module():
+    from tests.envinterp.conftest import set_variables
+
+    set_variables()
+
+
+def teardown_module():
+    from tests.envinterp.conftest import set_variables
+
+    set_variables()